+++ src/main/java/de/presti/ree6/sql/entities/custom/CustomEventAction.java
@@ -1,7 +1,7 @@
 package de.presti.ree6.sql.entities.custom;
 
 import com.google.gson.JsonElement;
-import de.presti.ree6.sql.converter.JsonAttributeConverter;
+import de.presti.ree6.sql.converter.JsonToBlobAttributeConverter;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -49,7 +49,7 @@ public class CustomEventAction {
      * The Actions that are to be executed.
      */
     @Column(name = "actions")
-    @Convert(converter = JsonAttributeConverter.class)
+    @Convert(converter = JsonToBlobAttributeConverter.class)
     JsonElement actions;
 
 }
